Project 9: Begin validation.
Added the start of the Validation hierarchy.  This is meant to be called by the Converter in between decoding and encoding.
This replaces the current notion of validation that was recently added.
No changes have been made to the Converter as of yet.  No acceptance tests have been created.  The next step is adding the measure validation.