Remove the usage of the checked exception NoSuchMethodException.
Created new RuntimeException, ConstructorNotFoundException to replace NoSuchMethodException near line 170 in Registry.java. If memory serves me then this was created to 'fail fast' when developing new handlers. It is not clear if handling this exception near line 137 hides this exception fail fast mechanism. Also revised the generics to reduce the wildcard use.